Added checks to File_transfer_detailed and Rclone_transfer to check for needed configs to run the IDRAC and WOL functionality.
Changed the default log level to info in the same
scripts.

Created the Plex_Heartbeat script to check if
there is Plex traffic on the network via an always
on proxy server. This includes IDRAC and WOL
logic.

Removed the Plex_power_off and Plex_power_on
scripts as they are no longer needed. The logic
for each is in the Plex_Heartbeat script.